+++ src/components/MovieList/MovieList.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 
 import { Paper, Typography } from '@material-ui/core';
 
-import Circular from '../Circular/Circular';
-import MovieListItem from './MovieListItem/MovieListItem';
+import Circular from '../Circular';
+import MovieListItem from './MovieListItem';
 
 import './movie-list.css';
